refactor(hf_adapter): we now have an IF that defines how to implement HF adapters for eval harness
